app/tables_docs/tasks.py
from langchain_core.tools import tool

@tool
def task_ref(arguments:str = None):
    """useful when you need meaning of column names in the Tasks table"""
    return f'''
    ---------------------------------------------------------------
    | TASK - (Activities)           | P6 EPPM Field / Column Name  |
    ---------------------------------------------------------------
    | act_end_date                  | Actual Finish                |
    | act_equip_qty                 | Actual Nonlabor Units        |
    | act_start_date                | Actual Start                 |
    | act_this_per_equip_qty        | Actual This Period Nonlabor Units |
    | act_this_per_work_qty         | Actual This Period Labor Units   |
    | act_work_qty                  | Actual Labor Units           |
    | auto_compute_act_flag         | Auto Compute Actuals         |
    | clndr_id                      | Calendar                     |
    | complete_pct_type             | Percent Complete Type        |
    | create_date                   | Added Date                   |
    | create_user                   | Added By                     |
    | cstr_date                     | Primary Constraint Date      |
    | cstr_date2                    | Secondary Constraint Date    |
    | cstr_type                     | Primary Constraint           |
    | cstr_type2                    | Secondary Constraint         |
    | driving_path_flag             | Longest Path                 |
    | duration_type                 | Duration Type                |
    | early_end_date                | Early Finish                 |
    | early_start_date              | Early Start                  |
    | est_wt                        | Est Weight (P6 Professional only) |
    | expect_end_date               | Expected Finish              |
    | external_early_start_date     | External Early Start         |
    | external_late_end_date        | External Late Finish         |
    | float_path                    | Float Path                   |
    | float_path_order              | Float Path Order             |
    | free_float_hr_cnt             | Free Float                   |
    | guid                          | Global Unique ID             |
    | late_end_date                 | Late Finish                  |
    | late_start_date               | Late Start                   |
    | location_id                   | Activity Location            |
    | lock_plan_flag                | Lock Remaining               |
    | phys_complete_pct             |                              |
    | priority_type                 | Activity Leveling Priority   |
    | proj_id                       | Project                      |
    | reend_date                    | Remaining Early Finish       |
    | rem_late_end_date             | Remaining Late Finish        |
    | rem_late_start_date           | Remaining Late Start         |
    | remain_drtn_hr_cnt            | Remaining Duration           |
    | remain_equip_qty              | Remaining Nonlabor Units     |
    | remain_work_qty               | Remaining Labor Units        |
    | restart_date                  | Remaining Early Start        |
    | resume_date                   | Resume Date                  |
    | rev_fdbk_flag                 | New Feedback                 |
    | review_end_date               | Review Finish (P6 Prof only) |
    | review_type                   | Review Status (P6 Prof only) |
    | rsrc_id                       | Primary Resource             |
    | status_code                   | Activity Status              |
    | suspend_date                  | Suspend Date                 |
    | target_drtn_hr_cnt            | Planned Duration (P6 EPPM) / Original or Planned Duration (P6 Prof) |
    | target_end_date               | Planned Finish               |
    | target_equip_qty              | Planned Nonlabor Units (P6 EPPM) / Budgeted or Planned Nonlabor Units (P6 Prof) |
    | target_start_date             | Planned Start                |
    | target_work_qty               | Planned Labor Units (P6 EPPM) / Budgeted or Planned Labor Units (P6 Prof) |
    | task_code                     | Activity ID                  |
    | task_id                       | Unique ID                    |
    | task_name                     | Activity Name                |
    | task_type                     | Activity Type                |
    | tmpl_guid                     | Methodology Global Unique ID |
    | total_float_hr_cnt            | Total Float                  |
    | update_date                   | Last Modified Date           |
    | update_user                   | Last Modified By             |
    | wbs_id                        | WBS                          |
    ---------------------------------------------------------------

  '''