requirements.txt
django
python-dotenv
djangorestframework
celery[redis]
django-redis
xer-reader
langchain
langchain-openai
langchain-google-genai
langgraph
tenacity
tiktoken
kombu
waitress
Pillow
redis
rq
django-rq
numpy
decord
opencv-python-headless